Move blockchain related stuff to blockchain_tests
