+++ src/blockchain/electrum.rs
@@ -172,6 +172,6 @@ impl ConfigurableBlockchain for ElectrumBlockchain {
 #[cfg(feature = "test-blockchains")]
 crate::bdk_blockchain_tests! {
     fn test_instance() -> ElectrumBlockchain {
-        ElectrumBlockchain::from(Client::new(&testutils::get_electrum_url()).unwrap())
+        ElectrumBlockchain::from(Client::new(&testutils::blockchain_tests::get_electrum_url()).unwrap())
     }
 }
